requirements.txt
torch
torchvision
torchaudio
pillow
matplotlib
accelerate